requirements.txt
azure-common==1.1.4
azure-nspkg==1.0.0
azure-storage==0.33.0
bottle==0.12.10
cffi==1.8.3
cryptography==1.5.2
freeze==1.0.10
idna==2.1
pyasn1==0.1.9
pycparser==2.16
PyJWT==1.4.2
Pykka==1.2.1
pymongo==3.3.0
python-dateutil==2.5.3
pytz==2016.7
requests==2.11.1
six==1.10.0
voluptuous
sendgrid==2.1.1